Authenticator: introduce an event-based API to let the BS know when the AuthClient db is modified.

+++ authenticator.rb
@@ -1,36 +1,38 @@
 class Authenticator < MObject
-  attr_reader :maclist, :config
+  attr_accessor :bs
 
-  def add_client(mac, interface, vlan, ipaddr=nil)
-    client = AuthClient.first_or_create(:macaddr => mac, :vlan => vlan,
-                                        :ipaddress => ipaddr, :interface => interface)
+  def add_client(mac, interface, vlan, ip=nil)
+    client = AuthClient.first_or_create(:macaddr   => mac,
+                                        :interface => interface,
+                                        :vlan      => vlan,
+                                        :ipaddress => ip)
     client.save
+    notify(:on_client_added, client)
   end
 
-  def add_ip_address(mac, ipaddr)
+  def update_client(mac, changes={})
+    client = AuthClient.get(mac)
+    return false if client.nil?
+    notify(:on_client_deleted, client)
+    client.update(changes)
+    notify(:on_client_added, client)
+    true
   end
 
   def del_client(mac)
     client = AuthClient.get(mac)
+    return false if client.nil?
+    notify(:on_client_deleted, client)
     client.destroy
+    true
   end
 
   def del_all_clients
     AuthClient.destroy
   end
 
-  def update_client(mac, uHash={})
-    client = AuthClient.get(mac)
-    client.update(uHash)
-  end
-
-  def get(mac)
+  def get_client(mac)
     AuthClient.get(mac)
-  end
-
-  def getIP(mac)
-    client = AuthClient.get(mac)
-    return client.nil? ? nil : client.ipaddress, client.vlan
   end
 
   def list_clients(interface, vlan=nil)
@@ -39,6 +41,14 @@ class Authenticator < MObject
       AuthClient.all(:vlan => vlan.to_s, :interface => interface)
     else
       AuthClient.all
+    end
+  end
+
+  private
+
+  def notify(event, *args)
+    if @bs && @bs.respond_to? event
+      @bs.call(event, *args)
     end
   end
